move useMetricsQuery  to ServicesView
